Move DOM injections to separate file
